move more modules to setup
